Allow customization at both page and component levels
