+++ packages/template-retail-react-app/app/pages/product-list/utils.js
@@ -1,0 +1,10 @@
+/*
+ * Copyright (c) 2024, Salesforce, Inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+import {renderPromoCallout as _render} from '@salesforce/retail-react-app/app/components/product-tile/utils'
+
+export const renderPromoCallout = _render
